separate fiels for each class
