+++ lib/webvtt.rb
@@ -11,4 +11,9 @@ module WebVTT
 end
 
 require "webvtt/parser"
+require "webvtt/blob"
+require "webvtt/file"
+require "webvtt/cue"
+require "webvtt/annotation"
+require "webvtt/timestamp"
 require "webvtt/segmenter"
